+++ veloxquant_mlx/cache/__init__.py
@@ -13,6 +13,7 @@ from veloxquant_mlx.cache.zipcache_cache import ZipCacheKVCache
 from veloxquant_mlx.cache.snapkv_cache import SnapKVKVCache
 from veloxquant_mlx.cache.streaming_llm_cache import StreamingLLMKVCache
 from veloxquant_mlx.cache.h2o_cache import H2OKVCache
+from veloxquant_mlx.cache.tova_cache import TOVAKVCache
 
 __all__ = [
     "KVCacheBuilder",
@@ -30,4 +31,5 @@ __all__ = [
     "SnapKVKVCache",
     "StreamingLLMKVCache",
     "H2OKVCache",
+    "TOVAKVCache",
 ]